+++ src/main/java/com/example/fintbackend/repositories/UserRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends CrudRepository<Users, Long> {
+public interface UserRepository extends CrudRepository<Users, Integer> {
     //TODO: This Users parameter is being called from 'models'.
     Optional<Users> findByEmail(String email);
 }
